Update pi_E_753_1CD.py
When unpowered (or in open loop mode) the z position is capable of drifting out of the legal range (for example -0.0123um). This can then result in an assertion error when calling a 'zero' relative move after power up or initialize (triggered by "assert self.z_min <= self.z <= self.z_max"). The new code block in the '__init__' steers the z just inside the legal range if it's outside (or hovering around the zero position which is typical).