Update pi_E_753_1CD.py
Actually 115200 is a valid board rate (as is 57600), with 115200 being the user manual default. The change from 115200 to 57600 seemed to fix what is now understood as a 'stochastic' failure on power up. Every few power cycles the controller will fail to response to the initial '*IDN?' call (return identity) which triggers and assertion error in the ._send() method (no response). Fortunately it seems to be robust if the initialize sequence is successful.