Update pi_E_753_1CD.py
Bug found that was causing the previously reported 'stochastic' failure on power up. The 'stochastic' nature came from whether or not the piezo had drifted outside the legal range, and if so it would attempt a 'MOV' command to legalize. However 'MOV' cannot be issued without a prior 'SVO' command given be _set_servo_enable(True) which has now been moved up in the __init__. The problem was made more cryptic by a secondary failure from the *IDN? call upon reboot which would fail if the error log was not cleared. A precautionary call to _check_errors() has now been added as the first call to help debug any future issues.